# server/swagger.yaml
openapi: 3.0.3
info:
  title: Volleyball Scoreboard API
  description: |
    API для управления табло волейбольного матча и системой авторизации.

    ## Провайдеры данных
    Система поддерживает два провайдера, переключаемых через переменную окружения `DB_PROVIDER`:
    - **firebase** — Firebase Firestore + Firebase Auth
    - **pocketbase** — PocketBase (self-hosted)

    ## Аутентификация
    1. Получите токен через `POST /api/auth/login`
    2. Используйте его в заголовке `Authorization: Bearer <token>` для защищённых эндпоинтов
    Все эндпоинты `/api/scoreboard/*` и `/api/matches/*` требуют авторизации.
    Только `/api/auth/login` и `/health` доступны без токена.

    ## Управление пользователями
    Создание, редактирование и удаление пользователей осуществляется **не через API**,
    а через админ-панель провайдера:
    - **Firebase Console** → Authentication → Users
    - **PocketBase Admin** → `http://your-server:8090/_/` → Collections → scoreusers
  version: 1.0.0
  contact:
    name: Volleyball Scoreboard
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
...
security:
  - BearerAuth: []
paths:
...
  /api/auth/login:
    post:
      summary: Войти в систему и получить токен
      description: |
        Аутентификация пользователя по username/email и паролю.
        Возвращает Bearer-токен для использования в заголовке `Authorization` последующих запросов.

        **Firebase:** email = `username@volleyball.local` (если передан только username)
        **PocketBase:** принимает username или email
      tags:
        - Auth
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - identity
                - password
              properties:
                identity:
                  type: string
                  description: Username или email
                  example: admin
                password:
                  type: string
                  format: password
                  description: Пароль пользователя
                  example: Admin@12345
      responses:
        '200':
          description: Успешный вход, получен токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Отсутствуют обязательные поля
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Неверные учётные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/me:
    get:
      summary: Информация о текущем пользователе
      description: Возвращает данные пользователя, идентифицированного по Bearer-токену
      tags:
        - Auth
      operationId: getCurrentUser
      responses:
        '200':
          description: Данные пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/token:
    post:
      summary: Информация о текущем токене
      description: Возвращает информацию о пользователе, связанном с переданным токеном
      tags:
        - Auth
      operationId: getTokenInfo
      responses:
        '200':
          description: Информация о токене
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenInfoResponse'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ===================== SCOREBOARD =====================
...
  /api/matches:
    get:
      summary: Получить список матчей
      description: Возвращает список завершённых матчей с пагинацией и фильтрами
      tags:
        - Matches
      operationId: listMatches
      parameters:
        - name: tournament
          in: query
          required: false
          schema:
            type: string
          description: Фильтр по названию турнира
        - name: game_type
          in: query
          required: false
          schema:
            type: string
            enum: [classic, beach]
          description: Тип матча
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
          description: Максимальное количество записей
      responses:
        '200':
          description: Список матчей
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Match'

    post:
      summary: Сохранить результат матча
      description: Записывает результат текущего матча в историю
      tags:
        - Matches
      operationId: saveMatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - game_id
              properties:
                game_id:
                  type: string
                  description: Идентификатор матча
                setHistory:
                  type: array
                  items:
                    type: object
                    properties:
                      home:
                        type: integer
                      away:
                        type: integer
                  description: История сетов (если не указан, берётся из табло)
                overallHome:
                  type: integer
                  description: Итоговый счёт домашней команды (сеты/фолы)
                overallAway:
                  type: integer
                  description: Итоговый счёт гостевой команды
      responses:
        '201':
          description: Результат сохранён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
        '400':
          description: game_id не указан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/matches/{id}:
    get:
      summary: Получить детали матча
      tags:
        - Matches
      operationId: getMatch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Идентификатор матча
      responses:
        '200':
          description: Детали матча
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
        '404':
          description: Матч не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer-токен, полученный через `POST /api/auth/login`.
        
        **Пример:**
        ```
        curl -X POST /api/auth/login -d '{"identity":"admin","password":"..."}'
        # → { "token": "eyJ..." }
        
        curl /api/auth/me -H "Authorization: Bearer eyJ..."
        ```
  schemas:
...
    CurrentUserResponse:
      type: object
      properties:
        user:
          type: object
          properties:
            uid:
              type: string
              example: abc123
            email:
              type: string
              example: admin@volleyball.local
            role:
              type: string
              enum: [admin, user]
              example: admin
            claims:
              type: object
              properties:
                role:
                  type: string
                admin:
                  type: boolean
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: Bearer-токен для заголовка Authorization
          example: eyJhbGciOiJSUzI1NiIsImtpZCI6Ii4uLiJ9...
        refreshToken:
          type: string
          description: Refresh token (для обновления access token)
          example: AOEToken...
        expiresIn:
          type: integer
          description: Время жизни токена в секундах
          example: 3600
        user:
          type: object
          properties:
            uid:
              type: string
            email:
              type: string
            username:
              type: string
            role:
              type: string
              enum: [admin, user]
            displayName:
              type: string
    TokenInfoResponse:
      type: object
      properties:
        message:
          type: string
          example: Используйте токен из заголовка Authorization
        user:
          type: object
          properties:
            uid:
              type: string
            email:
              type: string
            role:
              type: string

    # ---------- Scoreboard ----------
...
    Match:
      type: object
      properties:
        id:
          type: string
          description: Автоматически сгенерированный ID
        game_id:
          type: string
          example: game1
        date_time:
          type: string
          format: date-time
        home_team:
          type: string
          example: Динамо
        away_team:
          type: string
          example: Зенит
        tournament_name:
          type: string
          example: НВЛ
        overall_score:
          type: string
          example: "2:1"
        sets_score:
          type: array
          items:
            type: object
            properties:
              home:
                type: integer
              away:
                type: integer
        game_type:
          type: string
          enum: [classic, beach]
        two_wins_mode:
          type: boolean
        is_deleted:
          type: boolean
        deleted_at:
          type: string
          format: date-time
        notes:
          type: string

    # ---------- Error ----------
    Error:
      type: object
      properties:
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: game_id is required